ticketbookingapi/requirements.txt
alembic==1.18.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.10.1
PyMySQL==1.1.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5
qrcode[pil]>=8.0
flask-socketio>=5.5.0
Pillow>=10.0.0
python-socketio>=5.11.0
python-engineio>=4.11.0
simple-websocket>=1.0.0

# New dependencies for refactoring
python-dotenv>=1.0.0
marshmallow>=3.20.0
marshmallow-sqlalchemy>=1.0.0